Restructure for tree-shaking
